fix: normalize to 400g bags of tapioca flour
The tapioca flour bags we buy are 400g, so all quantities were
increased accordingly.

Fridging the mix made it easier to manage and helped raise the
outer-inter cooking speed ratio.

Fridging also meant we could easily portion the batter but cutting the
blocks in half, resulting in little cubes.

We found that shaping these into balls wasn't doing much and that the
squares are funnier anyway.